# gen/openapiv2/services.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "weewar/v1/models.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/gameinstances": {
      "get": {
        "summary": "ListGameInstances returns all available gameinstances",
        "operationId": "GameInstancesService_ListGameInstances",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListGameInstancesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "pagination.pageKey",
            "description": "*\nInstead of an offset an abstract  \"page\" key is provided that offers\nan opaque \"pointer\" into some offset in a result set.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pagination.pageOffset",
            "description": "*\nIf a pagekey is not supported we can also support a direct integer offset\nfor cases where it makes sense.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pagination.pageSize",
            "description": "*\nNumber of results to return.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "ownerId",
            "description": "May be filter by owner id",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "GameInstancesService"
        ]
      },
      "post": {
        "summary": "*\nCreate a new gameinstance",
        "operationId": "GameInstancesService_CreateGameInstance",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateGameInstanceResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateGameInstanceRequest"
            }
          }
        ],
        "tags": [
          "GameInstancesService"
        ]
      }
    },
    "/v1/gameinstances/{gameinstance.id}": {
      "patch": {
        "summary": "GetGameInstance returns a specific gameinstance with metadata",
        "operationId": "GameInstancesService_UpdateGameInstance",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateGameInstanceResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "gameinstance.id",
            "description": "Unique ID for the gameinstance",
            "in": "path",
            "required": true,
            "type": "string",
            "pattern": "[^/]+"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "gameinstance": {
                  "type": "object",
                  "properties": {
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "name": {
                      "type": "string",
                      "title": "Name if items have names"
                    },
                    "description": {
                      "type": "string",
                      "title": "Description if gameinstance has a description"
                    },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "title": "Some tags"
                    },
                    "imageUrl": {
                      "type": "string",
                      "title": "A possible image url"
                    },
                    "difficulty": {
                      "type": "string",
                      "title": "Difficulty - example attribute"
                    }
                  },
                  "title": "*\nGameInstance being updated"
                },
                "updateMask": {
                  "type": "string",
                  "description": "*\nMask of fields being updated in this GameInstance to make partial changes."
                }
              },
              "title": "UpdateGameInstanceRequest"
            }
          }
        ],
        "tags": [
          "GameInstancesService"
        ]
      }
    },
    "/v1/gameinstances/{id}": {
      "get": {
        "summary": "GetGameInstance returns a specific gameinstance with metadata",
        "operationId": "GameInstancesService_GetGameInstance",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetGameInstanceResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "version",
            "description": "Optional, defaults to default_version",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "GameInstancesService"
        ]
      },
      "delete": {
        "summary": "*\nDelete a particular gameinstance",
        "operationId": "GameInstancesService_DeleteGameInstance",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteGameInstanceResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "description": "*\nID of the gameinstance to be deleted.",
            "in": "path",
            "required": true,
            "type": "string",
            "pattern": "[^/]+"
          }
        ],
        "tags": [
          "GameInstancesService"
        ]
      }
    },
    "/v1/gameinstances:batchGet": {
      "get": {
        "summary": "*\nBatch get multiple gameinstances by ID",
        "operationId": "GameInstancesService_GetGameInstances",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetGameInstancesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "ids",
            "description": "*\nIDs of the gameinstance to be fetched",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "tags": [
          "GameInstancesService"
        ]
      }
    },
    "/v1/games": {
      "get": {
        "summary": "ListGames returns all available games",
        "operationId": "GamesService_ListGames",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListGamesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "pagination.pageKey",
            "description": "*\nInstead of an offset an abstract  \"page\" key is provided that offers\nan opaque \"pointer\" into some offset in a result set.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pagination.pageOffset",
            "description": "*\nIf a pagekey is not supported we can also support a direct integer offset\nfor cases where it makes sense.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pagination.pageSize",
            "description": "*\nNumber of results to return.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "ownerId",
            "description": "May be filter by owner id",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "GamesService"
        ]
      },
      "post": {
        "summary": "*\nCreate a new game",
        "operationId": "GamesService_CreateGame",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateGameResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateGameRequest"
            }
          }
        ],
        "tags": [
          "GamesService"
        ]
      }
    },
    "/v1/games/{game.id}": {
      "patch": {
        "summary": "GetGame returns a specific game with metadata",
        "operationId": "GamesService_UpdateGame",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateGameResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "game.id",
            "description": "Unique ID for the game",
            "in": "path",
            "required": true,
            "type": "string",
            "pattern": "[^/]+"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "game": {
                  "type": "object",
                  "properties": {
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "name": {
                      "type": "string",
                      "title": "Name if items have names"
                    },
                    "description": {
                      "type": "string",
                      "title": "Description if game has a description"
                    },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "title": "Some tags"
                    },
                    "imageUrl": {
                      "type": "string",
                      "title": "A possible image url"
                    },
                    "difficulty": {
                      "type": "string",
                      "title": "Difficulty - example attribute"
                    }
                  },
                  "title": "*\nGame being updated"
                },
                "updateMask": {
                  "type": "string",
                  "description": "*\nMask of fields being updated in this Game to make partial changes."
                }
              },
              "title": "UpdateGameRequest"
            }
          }
        ],
        "tags": [
          "GamesService"
        ]
      }
    },
    "/v1/games/{id}": {
      "get": {
        "summary": "GetGame returns a specific game with metadata",
        "operationId": "GamesService_GetGame",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetGameResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "version",
            "description": "Optional, defaults to default_version",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "GamesService"
        ]
      },
      "delete": {
        "summary": "*\nDelete a particular game",
        "operationId": "GamesService_DeleteGame",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1DeleteGameResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "description": "*\nID of the game to be deleted.",
            "in": "path",
            "required": true,
            "type": "string",
            "pattern": "[^/]+"
          }
        ],
        "tags": [
          "GamesService"
        ]
      }
    },
    "/v1/games:batchGet": {
      "get": {
        "summary": "*\nBatch get multiple games by ID",
        "operationId": "GamesService_GetGames",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetGamesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "ids",
            "description": "*\nIDs of the game to be fetched",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "tags": [
          "GamesService"
        ]
      }
    },
...
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1CreateGameInstanceRequest": {
      "type": "object",
      "properties": {
        "gameinstance": {
          "$ref": "#/definitions/v1GameInstance",
          "title": "*\nGameInstance being updated"
        }
      },
      "title": "*\nGameInstance creation request object"
    },
    "v1CreateGameInstanceResponse": {
      "type": "object",
      "properties": {
        "gameinstance": {
          "$ref": "#/definitions/v1GameInstance",
          "title": "*\nGameInstance being created"
        },
        "fieldErrors": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "description": "*\nError specific to a field if there are any errors."
        }
      },
      "description": "*\nResponse of an gameinstance creation."
    },
    "v1CreateGameRequest": {
      "type": "object",
      "properties": {
        "game": {
          "$ref": "#/definitions/v1Game",
          "title": "*\nGame being updated"
        }
      },
      "title": "*\nGame creation request object"
    },
    "v1CreateGameResponse": {
      "type": "object",
      "properties": {
        "game": {
          "$ref": "#/definitions/v1Game",
          "title": "*\nGame being created"
        },
        "fieldErrors": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "description": "*\nError specific to a field if there are any errors."
        }
      },
      "description": "*\nResponse of an game creation."
    },
...
    "v1DeleteGameInstanceResponse": {
      "type": "object",
      "title": "*\nGameInstance deletion response"
    },
    "v1DeleteGameResponse": {
      "type": "object",
      "title": "*\nGame deletion response"
    },
...
    "v1Game": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "type": "string",
          "title": "Unique ID for the game"
        },
        "name": {
          "type": "string",
          "title": "Name if items have names"
        },
        "description": {
          "type": "string",
          "title": "Description if game has a description"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Some tags"
        },
        "imageUrl": {
          "type": "string",
          "title": "A possible image url"
        },
        "difficulty": {
          "type": "string",
          "title": "Difficulty - example attribute"
        }
      }
    },
    "v1GameInstance": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "id": {
          "type": "string",
          "title": "Unique ID for the gameinstance"
        },
        "name": {
          "type": "string",
          "title": "Name if items have names"
        },
        "description": {
          "type": "string",
          "title": "Description if gameinstance has a description"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Some tags"
        },
        "imageUrl": {
          "type": "string",
          "title": "A possible image url"
        },
        "difficulty": {
          "type": "string",
          "title": "Difficulty - example attribute"
        }
      }
    },
    "v1GetGameInstanceResponse": {
      "type": "object",
      "properties": {
        "gameinstance": {
          "$ref": "#/definitions/v1GameInstance"
        }
      }
    },
    "v1GetGameInstancesResponse": {
      "type": "object",
      "properties": {
        "gameinstances": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/v1GameInstance"
          }
        }
      },
      "title": "*\nGameInstance batch-get response"
    },
    "v1GetGameResponse": {
      "type": "object",
      "properties": {
        "game": {
          "$ref": "#/definitions/v1Game"
        }
      }
    },
    "v1GetGamesResponse": {
      "type": "object",
      "properties": {
        "games": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/v1Game"
          }
        }
      },
      "title": "*\nGame batch-get response"
    },
...
    "v1ListGameInstancesResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1GameInstance"
          }
        },
        "pagination": {
          "$ref": "#/definitions/v1PaginationResponse"
        }
      }
    },
    "v1ListGamesResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Game"
          }
        },
        "pagination": {
          "$ref": "#/definitions/v1PaginationResponse"
        }
      }
    },
...
    "v1PaginationResponse": {
      "type": "object",
      "properties": {
        "nextPageKey": {
          "type": "string",
          "description": "*\nThe key/pointer string that subsequent List requests should pass to\ncontinue the pagination."
        },
        "nextPageOffset": {
          "type": "integer",
          "format": "int32",
          "title": "*\nAlso support an integer offset if possible"
        },
        "hasMore": {
          "type": "boolean",
          "description": "*\nWhether theere are more results."
        },
        "totalResults": {
          "type": "integer",
          "format": "int32",
          "description": "*\nTotal number of results."
        }
      }
    },
    "v1UpdateGameInstanceResponse": {
      "type": "object",
      "properties": {
        "gameinstance": {
          "$ref": "#/definitions/v1GameInstance",
          "title": "*\nGameInstance being updated"
        }
      },
      "description": "*\nThe request for (partially) updating an GameInstance.",
      "title": "UpdateGameInstanceResponse"
    },
    "v1UpdateGameResponse": {
      "type": "object",
      "properties": {
        "game": {
          "$ref": "#/definitions/v1Game",
          "title": "*\nGame being updated"
        }
      },
      "description": "*\nThe request for (partially) updating an Game.",
      "title": "UpdateGameResponse"
    },
...
  }
}